+++ CHANGELOG.md
@@ -1,6 +1,14 @@
-## 1.0.0 (2023-04-02)
+# Changelog
 
+## 2026.2.0 (2026-02-21)
+
+### Features
+
+- **Styling:** Added support for injecting custom CSS styles into cards and templates via the `style` property.
+- **Container:** Added `decluttering-container` class to the host element for easier CSS targeting.
+
+## 1.0.0 (2023-04-02)
 
 ### Bug Fixes
 
-* card broken with HA 2023.4 beta and up ([0ccd5b0](https://github.com/custom-cards/decluttering-card/commit/0ccd5b05a99202c80de21606df1b8e94ea9ee668)), closes [#63](https://github.com/custom-cards/decluttering-card/issues/63)
+- card broken with HA 2023.4 beta and up ([0ccd5b0](https://github.com/custom-cards/decluttering-card/commit/0ccd5b05a99202c80de21606df1b8e94ea9ee668)), closes [#63](https://github.com/custom-cards/decluttering-card/issues/63)